ui: move unlock animation to ui/grpahics/lock_animation
Decouple from workflow, so it can be used in Rust.
