+++ src/workflow/unlock_bip39.c
@@ -1,77 +1,30 @@
+// Copyright 2020 Shift Cryptosecurity AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #include "unlock_bip39.h"
 
 #include <hardfault.h>
 #include <keystore.h>
 #include <memory/memory.h>
-#include <screen.h>
 #include <ui/components/trinary_input_string.h>
-#include <ui/components/ui_images.h>
 #include <ui/graphics/lock_animation.h>
-#include <ui/ugui/ugui.h>
 #include <ui/workflow_stack.h>
 #include <util.h>
-
-#ifndef TESTING
-#include <hal_timer.h>
-#include <platform/driver_init.h>
-#endif
 
 #include "blocking.h"
 #include "get_mnemonic_passphrase.h"
 #include "workflow.h"
-
-#define TIMEOUT_TICK_PERIOD_MS 100
-
-#if !defined(TESTING)
-static struct timer_task _animation_timer_task = {0};
-static int _animation_current_frame = 0;
-
-/**
- * Displays frames of the lock animation
- * at a regular rate until it's finished.
- * Leaves the last frame on the screen.
- */
-static void _animation_timer_cb(const struct timer_task* const timer_task)
-{
-    (void)timer_task;
-    if (_animation_current_frame == LOCK_ANIMATION_N_FRAMES) {
-        /* End of the animation */
-        return;
-    }
-
-    /* Draw the frame. */
-    UG_ClearBuffer();
-    position_t pos = {.left = (SCREEN_WIDTH - LOCK_ANIMATION_FRAME_WIDTH) / 2,
-                      .top = (SCREEN_HEIGHT - LOCK_ANIMATION_FRAME_HEIGHT) / 2};
-    dimension_t dim = {.width = LOCK_ANIMATION_FRAME_WIDTH, .height = LOCK_ANIMATION_FRAME_HEIGHT};
-    in_buffer_t image = {.data = lock_animation_get_frame(_animation_current_frame),
-                         .len = LOCK_ANIMATION_FRAME_SIZE};
-    graphics_draw_image(&pos, &dim, &image);
-    UG_SendBuffer();
-    _animation_current_frame++;
-}
-
-/**
- * Sets up a timer animating a lock icon every 100ms
- */
-static void _start_animation_timer(void)
-{
-    _animation_timer_task.interval = TIMEOUT_TICK_PERIOD_MS;
-    _animation_timer_task.cb = _animation_timer_cb;
-    _animation_timer_task.mode = TIMER_TASK_REPEAT;
-    timer_stop(&TIMER_0);
-    timer_add_task(&TIMER_0, &_animation_timer_task);
-    _animation_current_frame = 0;
-    timer_start(&TIMER_0);
-}
-
-static void _stop_animation_timer(void)
-{
-    timer_stop(&TIMER_0);
-    timer_remove_task(&TIMER_0, &_animation_timer_task);
-    timer_start(&TIMER_0);
-}
-#endif
 
 typedef struct {
     /**
@@ -116,13 +69,9 @@ static void _unlock_bip39_spin(workflow_t* self)
 {
     unlock_bip39_data_t* data = (unlock_bip39_data_t*)self->data;
 
-#ifndef TESTING
-    _start_animation_timer();
-#endif
+    lock_animation_start();
     bool unlock_result = keystore_unlock_bip39(data->mnemonic_passphrase);
-#ifndef TESTING
-    _stop_animation_timer();
-#endif
+    lock_animation_stop();
 
     if (!unlock_result) {
         Abort("bip39 unlock failed");
